♻️  refactor routes and controllers to standardize naming conventions and improve code organization
