+++ src/controllers/ServicesController.ts
@@ -1,10 +1,10 @@
 import { NextFunction, Request, Response } from 'express'
 import { sendResponse } from '../utils/utils'
-import { ServicesService } from '../services/Services'
+import { ServicesService } from '../services/ServicesService'
 import { CustomError } from '../middleware/errorHandler'
 import { ServiceDto } from '../dtos/Services'
 import { ZodError } from 'zod'
-import { ToolsService } from '../services/Tools'
+import { ToolsService } from '../services/ToolsService'
 
 export class ServiceController {
   private servicesService: ServicesService
