+++ src/controllers/UserController.ts
@@ -1,7 +1,7 @@
 import { NextFunction, Request, Response } from 'express'
-import { UserService } from '../services/User'
+import { UserService } from '../services/UserService'
 import { CustomError } from '../middleware/errorHandler'
-import { UserPreferencesService } from '../services/UserPreferences'
+import { UserPreferencesService } from '../services/UserPreferencesService'
 import jwt from 'jsonwebtoken'
 import bcrypt from 'bcrypt'
 import { config } from 'dotenv'
