+++ src/routes/ServicesRoutes.ts
@@ -1,5 +1,5 @@
 import { Router } from 'express'
-import { ServiceController } from '../controllers/Services'
+import { ServiceController } from '../controllers/ServicesController'
 import { validationMiddleware } from '../middleware/validationMiddleware'
 import { ServiceDto } from '../dtos/Services'
 import { authenticateToken, authorizeRole } from '../middleware/validateUser'
